questionary, investment_scores
small changes made to questionary, instead of having six investment strategies I narrowed down to three in order to be able to explain it better based upon our etfs, as well as adjusting the risk scores to in order to be properly added to the correct portfolio.  I deleted some things that  are not being used such as time score and speculative. Please do not merge this until everyone agrees on changes.